refactor(nutrition): M-2a — extract leaf types from MealEntrySheet.swift (UI-004)
Audit M-2a: first phase of decomposing the 1104-line MealEntrySheet.swift.
This phase moves the 4 standalone leaf types out of the file. None depend
on MealEntrySheet's state — they're either pure data models, pure parsing,
or framework-glue UIViewControllerRepresentable wrappers.

Extracted:
- Models/OpenFoodFactsModels.swift (~95 lines): OFFSearchResponse,
  OFFProductResponse, OFFProduct, FoodProduct
- Parsing/NutritionLabelParser.swift (~120 lines): bilingual EN/HE
  nutrition label parser + ParsedNutritionLabel + language hint
- Camera/NutritionCameraSheet.swift (~46 lines): UIImagePickerController
  wrapper for label photo capture
- Camera/BarcodeScanner.swift (~165 lines): BarcodeScannerSheet +
  BarcodeScannerView + BarcodeScannerVC + AVCaptureMetadataOutput delegate

MealEntrySheet.swift drops 1104 → 695 lines (~37% reduction). Audit
asks for a deeper decomposition (per-tab views + ViewModel) — that
work is M-2b/c. M-2a is the safe, mechanical first phase.

Visibility bumped on ParsedNutritionLanguageHint, ParsedNutritionLabel,
NutritionLabelParser from private → internal (referenced by MealEntrySheet
from a different file now). Removed unused AVFoundation import from main
MealEntrySheet.swift after BarcodeScanner extraction.

Plan: docs/superpowers/plans/2026-04-19-m2-mealentrysheet-decomposition.md

Pure structural refactor. Build green.
