+++ FitTracker/Views/Nutrition/Camera/NutritionCameraSheet.swift
@@ -1,0 +1,46 @@
+// FitTracker/Views/Nutrition/Camera/NutritionCameraSheet.swift
+// Camera UI wrapper around UIImagePickerController for capturing nutrition labels.
+// Extracted from MealEntrySheet.swift in Audit M-2a (UI-004 decomposition).
+
+import SwiftUI
+#if canImport(UIKit)
+import UIKit
+#endif
+
+#if os(iOS)
+struct NutritionCameraSheet: UIViewControllerRepresentable {
+    let onCapture: (UIImage) -> Void
+
+    func makeCoordinator() -> Coordinator {
+        Coordinator(onCapture: onCapture)
+    }
+
+    func makeUIViewController(context: Context) -> UIImagePickerController {
+        let picker = UIImagePickerController()
+        picker.sourceType = .camera
+        picker.cameraCaptureMode = .photo
+        picker.delegate = context.coordinator
+        return picker
+    }
+
+    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) {}
+
+    final class Coordinator: NSObject, UINavigationControllerDelegate, UIImagePickerControllerDelegate {
+        let onCapture: (UIImage) -> Void
+
+        init(onCapture: @escaping (UIImage) -> Void) {
+            self.onCapture = onCapture
+        }
+
+        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
+            if let image = info[.originalImage] as? UIImage {
+                onCapture(image)
+            }
+        }
+
+        func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
+            picker.dismiss(animated: true)
+        }
+    }
+}
+#endif
