+++ FitTracker/Views/Nutrition/MealEntrySheet.swift
@@ -3,7 +3,6 @@
 // Three tabs: Manual entry, Template picker, Food search (text + barcode).
 
 import SwiftUI
-import AVFoundation
 import Vision
 #if canImport(PhotosUI)
 import PhotosUI
@@ -693,412 +692,3 @@ struct MealEntrySheet: View {
         }
     }
 }
-
-// ─────────────────────────────────────────────────────────
-// MARK: – OpenFoodFacts response models
-// ─────────────────────────────────────────────────────────
-
-// Audit UI-017: visibility bumped from `private` to internal so
-// FoodSearchService can decode these. Otherwise unchanged.
-struct OFFSearchResponse: Decodable {
-    var products: [OFFProduct]
-}
-
-struct OFFProductResponse: Decodable {
-    var product: OFFProduct?
-}
-
-struct OFFProduct: Decodable {
-    var product_name: String?
-
-    struct Nutriments: Decodable {
-        var energyKcal100g: Double?
-        var proteins100g:   Double?
-        var carbohydrates100g: Double?
-        var fat100g:        Double?
-
-        enum CodingKeys: String, CodingKey {
-            case energyKcal100g    = "energy-kcal_100g"
-            case proteins100g      = "proteins_100g"
-            case carbohydrates100g = "carbohydrates_100g"
-            case fat100g           = "fat_100g"
-        }
-    }
-    var nutriments: Nutriments?
-}
-
-// ─────────────────────────────────────────────────────────
-// MARK: – Parsed food product (view model)
-// ─────────────────────────────────────────────────────────
-
-// Audit UI-017: visibility bumped from `private` to internal so
-// FoodSearchService can produce these. Otherwise unchanged.
-struct FoodProduct: Identifiable {
-    let id = UUID()
-    var name:             String
-    var caloriesPer100g:  Double?
-    var proteinPer100g:   Double?
-    var carbsPer100g:     Double?
-    var fatPer100g:       Double?
-    var referenceGrams:   Double = 100
-    var source:           MealEntrySource = .search
-    var sourceDescription: String = "Open Food Facts"
-    var searchAliases:    [String] = []
-
-    init?(from raw: OFFProduct) {
-        name             = raw.product_name ?? ""
-        caloriesPer100g  = raw.nutriments?.energyKcal100g.flatMap    { $0 > 0 ? $0 : nil }
-        proteinPer100g   = raw.nutriments?.proteins100g.flatMap       { $0 > 0 ? $0 : nil }
-        carbsPer100g     = raw.nutriments?.carbohydrates100g.flatMap  { $0 > 0 ? $0 : nil }
-        fatPer100g       = raw.nutriments?.fat100g.flatMap            { $0 > 0 ? $0 : nil }
-        source = .barcode
-        sourceDescription = "Open Food Facts barcode or text search"
-        searchAliases = [name]
-        return  // always succeed — caller filters empty names in UI
-    }
-
-    init(
-        name: String,
-        caloriesPer100g: Double?,
-        proteinPer100g: Double?,
-        carbsPer100g: Double?,
-        fatPer100g: Double?,
-        aliases: [String]
-    ) {
-        self.name = name
-        self.caloriesPer100g = caloriesPer100g
-        self.proteinPer100g = proteinPer100g
-        self.carbsPer100g = carbsPer100g
-        self.fatPer100g = fatPer100g
-        self.referenceGrams = 100
-        self.source = .search
-        self.sourceDescription = "Built-in reference food"
-        self.searchAliases = aliases
-    }
-
-    static let referenceFoods: [FoodProduct] = [
-        .init(name: "White Rice / אורז לבן", caloriesPer100g: 130, proteinPer100g: 2.4, carbsPer100g: 28.2, fatPer100g: 0.3, aliases: ["white rice", "rice", "אורז", "אורז לבן"]),
-        .init(name: "Chicken Breast / חזה עוף", caloriesPer100g: 165, proteinPer100g: 31.0, carbsPer100g: 0, fatPer100g: 3.6, aliases: ["chicken breast", "chicken", "חזה עוף", "עוף"]),
-        .init(name: "Greek Yogurt / יוגורט יווני", caloriesPer100g: 97, proteinPer100g: 9.0, carbsPer100g: 3.9, fatPer100g: 5.0, aliases: ["greek yogurt", "יוגורט יווני", "יוגורט"]),
-        .init(name: "Oats / שיבולת שועל", caloriesPer100g: 389, proteinPer100g: 16.9, carbsPer100g: 66.3, fatPer100g: 6.9, aliases: ["oats", "oatmeal", "שיבולת שועל", "קוואקר"]),
-        .init(name: "Egg / ביצה", caloriesPer100g: 143, proteinPer100g: 12.6, carbsPer100g: 0.7, fatPer100g: 9.5, aliases: ["egg", "eggs", "ביצה", "ביצים"])
-    ]
-}
-
-private enum ParsedNutritionLanguageHint {
-    case english
-    case hebrew
-    case mixed
-}
-
-private struct ParsedNutritionLabel {
-    var referenceGrams: Double
-    var calories: Double?
-    var proteinG: Double?
-    var carbsG: Double?
-    var fatG: Double?
-    var nameHint: String?
-    var detectedLanguageHint: ParsedNutritionLanguageHint
-}
-
-private enum NutritionLabelParser {
-    static func parse(_ rawText: String, fallbackReferenceGrams: Double) -> ParsedNutritionLabel? {
-        let normalized = rawText
-            .replacingOccurrences(of: "\u{00A0}", with: " ")
-            .replacingOccurrences(of: ",", with: ".")
-            .lowercased()
-
-        let lines = normalized
-            .split(whereSeparator: \.isNewline)
-            .map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
-            .filter { !$0.isEmpty }
-
-        guard !lines.isEmpty else { return nil }
-
-        let referenceGrams = detectReferenceGrams(in: lines) ?? fallbackReferenceGrams
-        let calories = extractCalories(in: lines)
-        let protein = extractValue(in: lines, keywords: ["protein", "proteins", "חלבון", "חלבונים"])
-        let carbs = extractValue(in: lines, keywords: ["carbs", "carbohydrate", "carbohydrates", "פחמימה", "פחמימות"])
-        let fat = extractValue(in: lines, keywords: ["fat", "total fat", "שומן", "שומנים"])
-
-        guard calories != nil || protein != nil || carbs != nil || fat != nil else { return nil }
-
-        let languageHint: ParsedNutritionLanguageHint
-        if normalized.contains(where: { $0.unicodeScalars.contains(where: { $0.value >= 0x0590 && $0.value <= 0x05FF }) }) {
-            languageHint = normalized.contains("protein") || normalized.contains("fat") ? .mixed : .hebrew
-        } else {
-            languageHint = .english
-        }
-
-        return ParsedNutritionLabel(
-            referenceGrams: referenceGrams,
-            calories: calories,
-            proteinG: protein,
-            carbsG: carbs,
-            fatG: fat,
-            nameHint: lines.first(where: { !$0.contains("calories") && !$0.contains("חלבון") && !$0.contains("פחמ") && !$0.contains("שומן") }),
-            detectedLanguageHint: languageHint
-        )
-    }
-
-    private static func detectReferenceGrams(in lines: [String]) -> Double? {
-        if lines.contains(where: { $0.contains("100g") || $0.contains("100 g") || $0.contains("100גרם") || $0.contains("100 גרם") || $0.contains("ל-100") }) {
-            return 100
-        }
-
-        let referenceKeywords = ["serving", "serving size", "מנה", "כמות למנה", "portion"]
-        for line in lines where referenceKeywords.contains(where: { line.contains($0) }) {
-            if let grams = extractGrams(from: line) {
-                return grams
-            }
-        }
-        return nil
-    }
-
-    private static func extractCalories(in lines: [String]) -> Double? {
-        for line in lines where ["calories", "energy", "kcal", "קלוריות", "אנרגיה"].contains(where: { line.contains($0) }) {
-            let values = extractNumbers(from: line)
-            if line.contains("kcal"), let kcal = values.last(where: { $0 < 1200 }) {
-                return kcal
-            }
-            if let single = values.first {
-                return single
-            }
-        }
-        return nil
-    }
-
-    private static func extractValue(in lines: [String], keywords: [String]) -> Double? {
-        for line in lines where keywords.contains(where: { line.contains($0) }) {
-            if let value = extractNumbers(from: line).first {
-                return value
-            }
-        }
-        return nil
-    }
-
-    private static func extractGrams(from line: String) -> Double? {
-        let pattern = #"(\d+(?:\.\d+)?)\s*(?:g|gr|gram|grams|גרם)"#
-        guard let regex = try? NSRegularExpression(pattern: pattern, options: []) else { return nil }
-        let nsRange = NSRange(line.startIndex..<line.endIndex, in: line)
-        guard let match = regex.firstMatch(in: line, options: [], range: nsRange),
-              let range = Range(match.range(at: 1), in: line) else { return nil }
-        return Double(String(line[range]))
-    }
-
-    private static func extractNumbers(from line: String) -> [Double] {
-        let pattern = #"\d+(?:\.\d+)?"#
-        guard let regex = try? NSRegularExpression(pattern: pattern, options: []) else { return [] }
-        let nsRange = NSRange(line.startIndex..<line.endIndex, in: line)
-        return regex.matches(in: line, options: [], range: nsRange).compactMap { match in
-            guard let range = Range(match.range(at: 0), in: line) else { return nil }
-            return Double(String(line[range]))
-        }
-    }
-}
-
-// ─────────────────────────────────────────────────────────
-// MARK: – Barcode Scanner (iOS only)
-// ─────────────────────────────────────────────────────────
-
-#if os(iOS)
-struct NutritionCameraSheet: UIViewControllerRepresentable {
-    let onCapture: (UIImage) -> Void
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator(onCapture: onCapture)
-    }
-
-    func makeUIViewController(context: Context) -> UIImagePickerController {
-        let picker = UIImagePickerController()
-        picker.sourceType = .camera
-        picker.cameraCaptureMode = .photo
-        picker.delegate = context.coordinator
-        return picker
-    }
-
-    func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) {}
-
-    final class Coordinator: NSObject, UINavigationControllerDelegate, UIImagePickerControllerDelegate {
-        let onCapture: (UIImage) -> Void
-
-        init(onCapture: @escaping (UIImage) -> Void) {
-            self.onCapture = onCapture
-        }
-
-        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-            if let image = info[.originalImage] as? UIImage {
-                onCapture(image)
-            }
-        }
-
-        func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
-            picker.dismiss(animated: true)
-        }
-    }
-}
-
-struct BarcodeScannerSheet: View {
-    let onScan: (String) -> Void
-    @Environment(\.dismiss) var dismiss
-
-    var body: some View {
-        NavigationStack {
-            BarcodeScannerView(onScan: { barcode in
-                onScan(barcode)
-            })
-            .ignoresSafeArea()
-            .navigationTitle("Scan Barcode")
-            .navigationBarTitleDisplayMode(.inline)
-            .toolbar {
-                ToolbarItem(placement: .cancellationAction) {
-                    Button("Cancel") { dismiss() }
-                }
-            }
-        }
-    }
-}
-
-struct BarcodeScannerView: UIViewControllerRepresentable {
-    let onScan: (String) -> Void
-
-    func makeCoordinator() -> Coordinator {
-        Coordinator(onScan: onScan)
-    }
-
-    func makeUIViewController(context: Context) -> BarcodeScannerVC {
-        let vc = BarcodeScannerVC()
-        vc.delegate = context.coordinator
-        return vc
-    }
-
-    func updateUIViewController(_ uiViewController: BarcodeScannerVC, context: Context) {}
-
-    // ── Coordinator ──────────────────────────────────────
-
-    final class Coordinator: NSObject, BarcodeScannerVCDelegate {
-        let onScan: (String) -> Void
-        private var didScan = false
-
-        init(onScan: @escaping (String) -> Void) { self.onScan = onScan }
-
-        func barcodeScannerVC(_ vc: BarcodeScannerVC, didScanBarcode barcode: String) {
-            guard !didScan else { return }
-            didScan = true
-            DispatchQueue.main.async { self.onScan(barcode) }
-        }
-    }
-}
-
-// ── Protocol ─────────────────────────────────────────────
-
-protocol BarcodeScannerVCDelegate: AnyObject {
-    func barcodeScannerVC(_ vc: BarcodeScannerVC, didScanBarcode barcode: String)
-}
-
-// ── UIViewController wrapping AVCaptureSession ────────────
-
-final class BarcodeScannerVC: UIViewController {
-    weak var delegate: BarcodeScannerVCDelegate?
-
-    private let session        = AVCaptureSession()
-    private var previewLayer:    AVCaptureVideoPreviewLayer?
-    private let metadataQueue  = DispatchQueue(label: "com.fittracker.barcode")
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .black
-        setupSession()
-    }
-
-    override func viewDidLayoutSubviews() {
-        super.viewDidLayoutSubviews()
-        previewLayer?.frame = view.bounds
-    }
-
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        metadataQueue.async {
-            if !self.session.isRunning { self.session.startRunning() }
-        }
-    }
-
-    override func viewWillDisappear(_ animated: Bool) {
-        super.viewWillDisappear(animated)
-        metadataQueue.async {
-            if self.session.isRunning { self.session.stopRunning() }
-        }
-    }
-
-    private func setupSession() {
-        guard AVCaptureDevice.authorizationStatus(for: .video) != .denied else {
-            showPermissionDenied()
-            return
-        }
-
-        session.beginConfiguration()
-        defer { session.commitConfiguration() }
-
-        guard let device = AVCaptureDevice.default(for: .video),
-              let input  = try? AVCaptureDeviceInput(device: device) else { return }
-
-        if session.canAddInput(input) { session.addInput(input) }
-
-        let output = AVCaptureMetadataOutput()
-        if session.canAddOutput(output) {
-            session.addOutput(output)
-            output.setMetadataObjectsDelegate(self, queue: metadataQueue)
-            let supported: [AVMetadataObject.ObjectType] = [.ean8, .ean13, .upce, .code128, .qr]
-            output.metadataObjectTypes = supported.filter { output.availableMetadataObjectTypes.contains($0) }
-        }
-
-        let preview = AVCaptureVideoPreviewLayer(session: session)
-        preview.videoGravity = .resizeAspectFill
-        preview.frame = view.bounds
-        view.layer.insertSublayer(preview, at: 0)
-        previewLayer = preview
-
-        // Scanning guide overlay
-        let guide = UIView()
-        guide.layer.borderColor = UIColor.white.withAlphaComponent(0.8).cgColor
-        guide.layer.borderWidth = 2
-        guide.layer.cornerRadius = 8
-        guide.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(guide)
-        NSLayoutConstraint.activate([
-            guide.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-            guide.centerYAnchor.constraint(equalTo: view.centerYAnchor),
-            guide.widthAnchor.constraint(equalTo: view.widthAnchor, multiplier: 0.7),
-            guide.heightAnchor.constraint(equalToConstant: 120)
-        ])
-    }
-
-    private func showPermissionDenied() {
-        DispatchQueue.main.async {
-            let label = UILabel()
-            label.text = "Camera access is required to scan barcodes.\nPlease enable it in Settings."
-            label.textColor = .white
-            label.numberOfLines = 0
-            label.textAlignment = .center
-            label.translatesAutoresizingMaskIntoConstraints = false
-            self.view.addSubview(label)
-            NSLayoutConstraint.activate([
-                label.centerXAnchor.constraint(equalTo: self.view.centerXAnchor),
-                label.centerYAnchor.constraint(equalTo: self.view.centerYAnchor),
-                label.widthAnchor.constraint(equalTo: self.view.widthAnchor, constant: -40)
-            ])
-        }
-    }
-}
-
-extension BarcodeScannerVC: AVCaptureMetadataOutputObjectsDelegate {
-    func metadataOutput(_ output: AVCaptureMetadataOutput,
-                        didOutput metadataObjects: [AVMetadataObject],
-                        from connection: AVCaptureConnection) {
-        guard let obj = metadataObjects.first as? AVMetadataMachineReadableCodeObject,
-              let value = obj.stringValue else { return }
-        session.stopRunning()
-        delegate?.barcodeScannerVC(self, didScanBarcode: value)
-    }
-}
-#endif
